Implement PR suggestions, remove lifetime from Options
