+++ examples/custom_href_resolver.rs
@@ -1,41 +1,30 @@
 use std::sync::Arc;
 
-use usvg::{ImageHrefResolver, ImageKind};
+use usvg::{ImageHrefResolver, ImageKind, OptionsRef};
 
 fn main() {
-    let args: Vec<String> = std::env::args().collect();
-    if args.len() != 3 {
-        println!("Usage:\n\tminimal <in-svg> <out-png>");
-        return;
-    }
-
     let mut opt = usvg::Options::default();
-    // Get file's absolute directory.
-    opt.resources_dir = std::fs::canonicalize(&args[1])
-        .ok()
-        .and_then(|p| p.parent().map(|p| p.to_path_buf()));
-    opt.fontdb.load_system_fonts();
 
-    // We know that our SVG won't have DataUrl hrefs.
-    let resolve_data = Box::new(|_: &str, _: Arc<Vec<u8>>| None);
+    let ferris_image_path = "./examples/ferris.png";
+    let ferris_image = Arc::new(std::fs::read(ferris_image_path).unwrap());
 
-    // Let's treat href as URL and download PNG image via network.
-    let resolve_string = Box::new(|href: &str| {
-        reqwest::blocking::get(href)
-            .and_then(|data| data.bytes())
-            .map(|bytes| bytes.to_vec())
-            .ok()
-            .map(|data: Vec<u8>| ImageKind::PNG(Arc::new(data)))
+    // We know that our SVG won't have DataUrl hrefs, just return None for such case.
+    let resolve_data = Box::new(|_: &str, _: Arc<Vec<u8>>, _: &OptionsRef| None);
+
+    // Here we handle xlink:href attribute as string, let's use already loaded Ferris image to match that string.
+    let resolve_string = Box::new(move |href: &str, _: &OptionsRef| match href {
+        "ferris_image" => Some(ImageKind::PNG(ferris_image.clone())),
+        _ => None
     });
 
     // Assign new ImageHrefResolver option using our closures.
-    // Now `href` strings in `<image>` elements will be treated as URLs to PNG images.
     opt.image_href_resolver = Some(ImageHrefResolver {
         resolve_data,
         resolve_string,
     });
 
-    let svg_data = std::fs::read(&args[1]).unwrap();
+    let svg_with_ferris_path = "./examples/custom_href_resolver.svg";
+    let svg_data = std::fs::read(svg_with_ferris_path).unwrap();
     let rtree = usvg::Tree::from_data(&svg_data, &opt.to_ref()).unwrap();
 
     let pixmap_size = rtree.svg_node().size.to_screen_size();
@@ -49,5 +38,5 @@ fn main() {
     )
     .unwrap();
 
-    pixmap.save_png(&args[2]).unwrap();
+    pixmap.save_png("custom_href_resolver.png").unwrap();
 }
