+++ usvg/src/lib.rs
@@ -112,7 +112,6 @@ mod error;
 #[cfg(feature = "filter")] pub mod filter;
 mod geom;
 mod image;
-mod image_href_resolver;
 mod marker;
 mod mask;
 mod numbers;
@@ -127,8 +126,8 @@ mod switch;
 mod units;
 mod use_node;
 
-pub use image_href_resolver::ImageHrefResolver;
-pub use image_href_resolver::default_resolver::{create_default_data_resolver, create_default_string_resolver};
+pub use image::ImageHrefResolver;
+pub use image::{create_default_data_resolver, create_default_string_resolver};
 pub use svgtypes::{Align, AspectRatio};
 
 #[cfg(feature = "text")] pub use fontdb;
